fix: add explicit casts to fix sign conversion warnings
- Cast bytes to uint32_t before bit shifting in manifest_generate_tests.cpp
- Fix cast placement in read_le16() in manifest.cpp

+++ src/manifest/manifest.cpp
@@ -43,7 +43,7 @@ uint32_t read_le32(const uint8_t* p) {
 }
 
 uint16_t read_le16(const uint8_t* p) {
-    return static_cast<uint16_t>(p[0]) | static_cast<uint16_t>(p[1] << 8);
+    return static_cast<uint16_t>(p[0]) | static_cast<uint16_t>(static_cast<uint16_t>(p[1]) << 8);
 }
 
 ManifestHeader parse_header(const uint8_t* data) {
